[fixed] change varable type to double - Nov 11, 2022

+++ fcu_cs/22fall_Data_Structures/DS_HW2-Arithmetic/Infix2Postfix_Arithmetic.cpp
@@ -87,7 +87,7 @@ bool checkBrackets(){ // check brackets are valid
 
 
 void postfix_operator(string str){
-    stack <int> nums; // store the numbers
+    stack <double> nums; // store the numbers
 
     for(int i=0; i<str.length(); i++){
         if(isDigit(str[i])){ // if the character is a number
@@ -100,8 +100,8 @@ void postfix_operator(string str){
             nums.push(val);
         }
         else if(str[i] != ' '){ // if the character is an operator
-            int b = nums.top(); nums.pop(); // pop the top two numbers
-            int a = nums.top(); nums.pop();
+            double b = nums.top(); nums.pop(); // pop the top two numbers
+            double a = nums.top(); nums.pop();
 
             switch(str[i]){
                 case '+':
